requirements.txt
streamlit==1.30.0
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0
scipy==1.12.0
